Feat: Divide views into separate files in 'user' app
