+++ diningcoach/user/urls.py
@@ -1,21 +1,12 @@
 from django.urls import path, include
 
-from user.views import (
-  UserSignUpView, UserLoginView, UserLogoutView, CustomizedTokenRefreshView,
-  KakaoSignInView, GoogleSignInView, AppleSignInView, NaverSignInView,
-  UserProfileView, UserHealthView, ConsentTermsView
-)
+from user.views.account_views import UserSignUpView, UserLoginView, UserLogoutView, CustomizedTokenRefreshView
+from user.views.platform_views import KakaoSignInView, GoogleSignInView, AppleSignInView, NaverSignInView
+from user.views.info_views import UserProfileView, UserHealthView, ConsentTermsView
 
-from rest_framework_simplejwt.views import (
-  TokenObtainPairView, TokenRefreshView, TokenVerifyView
-)
+from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView, TokenVerifyView
 
 urlpatterns = [
-  path('auth/', include('dj_rest_auth.urls')),
-  path('auth/registration/', include('dj_rest_auth.registration.urls')),
-  path('allauth/', include('allauth.urls')),
-  # path('allauth/account/', include('allauth.account.urls')),
-
   path('account/signup/', UserSignUpView.as_view()),
   path('account/login/', UserLoginView.as_view()),
   path('account/logout/', UserLogoutView.as_view()),
@@ -26,13 +17,18 @@ urlpatterns = [
   path('platform/apple/', AppleSignInView.as_view()),
   path('platform/naver/', NaverSignInView.as_view()),
 
+  path('info/profile/', UserProfileView.as_view()),
+  path('info/health/', UserHealthView.as_view()),
+  path('info/consent/', ConsentTermsView.as_view()),
+
   path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
   path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
   path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
 
-  path('info/profile/', UserProfileView.as_view()),
-  path('info/health/', UserHealthView.as_view()),
-  path('info/consent/', ConsentTermsView.as_view()),
+  path('auth/', include('dj_rest_auth.urls')),
+  path('auth/registration/', include('dj_rest_auth.registration.urls')),
+  path('allauth/', include('allauth.urls')),
+  # path('allauth/account/', include('allauth.account.urls')),
 ]
 
 # dj-rest-auth
